removed robocup functions from usbd_cdc_if.c
